migrate from postgres to sqlite3
add a genserver that keeps the old database contents
in memory, and combines it with the new database
when serving requests.

I actually prefer this as it means I can always export
the database to a file and have a backup of the entire
database, and can modify the JSON file directly if I
want to.